backend/dama/requirements.txt
asgiref==3.8.1
des==1.0.6
Django==5.1.3
django-cors-headers==4.6.0
django-filter==25.1
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pycryptodomex==3.21.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.14.1
tzdata==2024.2
uritemplate==4.1.1
